refactor: address PR #735 review feedback
- Extract shared LegalityUi static helper (GetStatus, GetFirstIssue,
  GetStatusColor, GetStatusIcon, GetStatusLabel) so the Report tab,
  Import preview, storage component, and slot component all classify
  and render tri-state status identically. Removes the "keep in sync"
  comments that were a maintenance hazard.
- ShowdownImportDialog.ParseTextAsync now wraps the conversion +
  analysis loop in try/finally so isParsing always clears and the
  dialog can't get stuck on "Analyzing…" if an exception escapes.
- ImportAsync caches the SaveFile reference used at parse time and
  bails with a "please parse again" snackbar if the current save
  differs — prevents importing PKMs generated against a stale save.
