+++ Pkmds.Rcl/Components/Dialogs/ShowdownImportDialog.razor.cs
@@ -1,5 +1,3 @@
-using PKHexSeverity = PKHeX.Core.Severity;
-
 namespace Pkmds.Rcl.Components.Dialogs;
 
 public partial class ShowdownImportDialog
@@ -11,6 +9,7 @@ public partial class ShowdownImportDialog
     private bool isFetching;
     private bool isParsing;
     private List<ParsedEntry> parsedEntries = [];
+    private SaveFile? parsedWithSaveFile;
     private string? pasteInfo;
 
     [CascadingParameter]
@@ -30,29 +29,40 @@ public partial class ShowdownImportDialog
 
         isParsing = true;
         parsedEntries = [];
+        // Record which save the entries are valid for so Import can detect a mismatch
+        // (e.g. the save was swapped between parse and import).
+        parsedWithSaveFile = AppState.SaveFile;
         StateHasChanged();
 
-        // Yield so the spinner paints before per-set conversion + legality analysis runs.
-        // Task.Delay(1) (not Task.Yield) because in Blazor WASM, Yield stays on the same
-        // JS macrotask and never lets the browser paint.
-        await Task.Delay(1);
-
-        var sets = AppService.ParseShowdownText(inputText);
-        var entries = new List<ParsedEntry>(sets.Count);
-        foreach (var set in sets)
+        try
         {
-            var pkm = AppService.ConvertShowdownSetToPkm(set);
-            var (status, firstIssue) = AnalyzeLegality(pkm);
-            entries.Add(new ParsedEntry(set, pkm, status, firstIssue));
-
-            // Inter-entry yield so the browser can process input between conversions,
-            // matching the pattern used by LegalityReportTab's batch legalize sweep.
+            // Yield so the spinner paints before per-set conversion + legality analysis runs.
+            // Task.Delay(1) (not Task.Yield) because in Blazor WASM, Yield stays on the same
+            // JS macrotask and never lets the browser paint.
             await Task.Delay(1);
-        }
 
-        parsedEntries = entries;
-        isParsing = false;
-        StateHasChanged();
+            var sets = AppService.ParseShowdownText(inputText);
+            var entries = new List<ParsedEntry>(sets.Count);
+            foreach (var set in sets)
+            {
+                var pkm = AppService.ConvertShowdownSetToPkm(set);
+                var (status, firstIssue) = AnalyzeLegality(pkm);
+                entries.Add(new ParsedEntry(set, pkm, status, firstIssue));
+
+                // Inter-entry yield so the browser can process input between conversions,
+                // matching the pattern used by LegalityReportTab's batch legalize sweep.
+                await Task.Delay(1);
+            }
+
+            parsedEntries = entries;
+        }
+        finally
+        {
+            // Always clear the spinner / re-enable inputs, even if conversion or analysis
+            // threw — otherwise the dialog can get stuck in "Analyzing…" forever.
+            isParsing = false;
+            StateHasChanged();
+        }
     }
 
     private (LegalityStatus? Status, string FirstIssue) AnalyzeLegality(PKM? pkm)
@@ -63,98 +73,17 @@ public partial class ShowdownImportDialog
         }
 
         var la = AppService.GetLegalityAnalysis(pkm);
-        var status = GetStatus(la);
+        var status = LegalityUi.GetStatus(la);
         var firstIssue = status == LegalityStatus.Legal
             ? string.Empty
-            : GetFirstIssue(la);
+            : LegalityUi.GetFirstIssue(la);
         return (status, firstIssue);
     }
 
-    // Mirrors LegalityReportTab.GetStatus — keep these in sync.
-    private static LegalityStatus GetStatus(LegalityAnalysis la)
-    {
-        var hasInvalid = la.Results.Any(r => r.Judgement == PKHexSeverity.Invalid)
-                         || !MoveResult.AllValid(la.Info.Moves)
-                         || !MoveResult.AllValid(la.Info.Relearn);
-
-        if (hasInvalid)
-        {
-            return LegalityStatus.Illegal;
-        }
-
-        var hasFishy = la.Results.Any(r => r.Judgement == PKHexSeverity.Fishy);
-        return hasFishy
-            ? LegalityStatus.Fishy
-            : LegalityStatus.Legal;
-    }
-
-    // Mirrors LegalityReportTab.GetFirstIssue — keep these in sync.
-    private static string GetFirstIssue(LegalityAnalysis la)
-    {
-        var ctx = LegalityLocalizationContext.Create(la);
-
-        // Prefer Invalid over Fishy so the more severe issue wins when both are present.
-        // CheckResult.Valid is true for Fishy judgements, so match on Judgement directly.
-        foreach (var result in la.Results)
-        {
-            if (result.Judgement == PKHexSeverity.Invalid)
-            {
-                return ctx.Humanize(in result);
-            }
-        }
-
-        if (!MoveResult.AllValid(la.Info.Moves))
-        {
-            return "Invalid move detected.";
-        }
-
-        if (!MoveResult.AllValid(la.Info.Relearn))
-        {
-            return "Invalid relearn move detected.";
-        }
-
-        foreach (var result in la.Results)
-        {
-            if (result.Judgement == PKHexSeverity.Fishy)
-            {
-                return ctx.Humanize(in result);
-            }
-        }
-
-        return string.Empty;
-    }
-
-    private static Color GetStatusColor(LegalityStatus status) => status switch
-    {
-        LegalityStatus.Legal => Color.Success,
-        LegalityStatus.Fishy => Color.Warning,
-        LegalityStatus.Illegal => Color.Error,
-        _ => Color.Default
-    };
-
-    private static string GetStatusIcon(LegalityStatus status) => status switch
-    {
-        LegalityStatus.Legal => Icons.Material.Filled.CheckCircle,
-        LegalityStatus.Fishy => Icons.Material.Filled.Warning,
-        LegalityStatus.Illegal => Icons.Material.Filled.Cancel,
-        _ => Icons.Material.Filled.Help
-    };
-
-    private static string GetStatusTooltip(ParsedEntry entry) => entry.Status switch
-    {
-        LegalityStatus.Legal => "Legal",
-        _ => string.IsNullOrEmpty(entry.FirstIssue)
-            ? GetStatusLabel(entry.Status)
-            : entry.FirstIssue
-    };
-
-    private static string GetStatusLabel(LegalityStatus? status) => status switch
-    {
-        LegalityStatus.Legal => "Legal",
-        LegalityStatus.Fishy => "Fishy",
-        LegalityStatus.Illegal => "Illegal",
-        _ => "Unknown"
-    };
+    private static string GetStatusTooltip(ParsedEntry entry) =>
+        string.IsNullOrEmpty(entry.FirstIssue)
+            ? LegalityUi.GetStatusLabel(entry.Status ?? LegalityStatus.Legal)
+            : entry.FirstIssue;
 
     private async Task FetchUrlAsync()
     {
@@ -257,6 +186,18 @@ public partial class ShowdownImportDialog
         if (AppState.SaveFile is not { } sav)
         {
             Snackbar.Add("No save file loaded.", Severity.Warning);
+            return;
+        }
+
+        // Defensive: the dialog is modal so the save shouldn't change mid-flight, but if
+        // it somehow did (or the user parsed before loading a save), the stored PKMs
+        // belong to a different target and must not be imported.
+        if (!ReferenceEquals(parsedWithSaveFile, sav))
+        {
+            Snackbar.Add("Save file changed since parsing. Please parse again.", Severity.Warning);
+            parsedEntries = [];
+            parsedWithSaveFile = null;
+            StateHasChanged();
             return;
         }
